Read boundary regions as well

+++ yt/geometry/ramses_oct_container.pyx
@@ -94,7 +94,7 @@ cdef class RAMSESOctreeContainer(SparseOctreeContainer):
                    np.uint8_t[:] levels,
                    np.uint8_t[:] cell_inds,
                    np.int64_t[:] file_inds,
-                   np.int64_t[:] domains,
+                   np.int32_t[:] domains,
                    dict dest_fields, dict source_fields,
                    np.int32_t domain,
                    np.int64_t offset = 0
